Fix release build: make template implementation as inline
